:art: folder refactor
