+++ src/components/icon/Dimensions.tsx
@@ -1,5 +1,5 @@
 import IconContainer from "./IconContainer";
-import dimensions from "../../../assets/icons/dimensions.png";
+import dimensions from "../../assets/icons/dimensions.png";
 
 const DimensionsIcon = () => {
   return <IconContainer imgUrl={dimensions} label="Dimensions" />;
