+++ src/components/icon/SpatialIcon.tsx
@@ -1,5 +1,5 @@
 import IconContainer from "./IconContainer";
-import spatial from "../../../assets/icons/spatial.png";
+import spatial from "../../assets/icons/spatial.png";
 
 const SpatialIcon = () => {
   return <IconContainer imgUrl={spatial} label="Spatial" />;
