+++ src/components/icon/TimeRangeIcon.tsx
@@ -1,5 +1,5 @@
 import IconContainer from "./IconContainer";
-import timeRange from "../../../assets/logos/time-range.png";
+import timeRange from "../../assets/logos/time-range.png";
 
 const TimeRangeIcon = () => {
   return <IconContainer imgUrl={timeRange} label="Time Range" />;
